feat: Moving Slack notifier from EE to community
closes gravitee-io/issues#7752
