+++ src/main/resources/plugin.properties
@@ -2,6 +2,6 @@ id=slack-notifier
 name=Slack
 version=${project.version}
 description=${project.description}
-class=com.graviteesource.notifier.slack.SlackNotifier
+class=io.gravitee.notifier.slack.SlackNotifier
 type=notifier
 icon=slack.svg
