+++ src/main/resources/schemas/schema-form.json
@@ -1,6 +1,6 @@
 {
   "type" : "object",
-  "id" : "urn:jsonschema:com:graviteesource:notifier:slack:configuration:SlackNotifierConfiguration",
+  "id" : "urn:jsonschema:io:gravitee:notifier:slack:configuration:SlackNotifierConfiguration",
   "properties" : {
     "channel": {
       "title": "Slack channel",
